css/selectors/focus-visible-029.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8" />
  <title>CSS Test (Selectors): :focus-visible does not match after a fragment navigation using a pointing device</title>
  <link rel="author" title="Harry Moore" href="harry-moore@live.co.uk" />
  <link rel="help" href="https://drafts.csswg.org/selectors-4/#the-focus-visible-pseudo" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <style>
    @supports not selector(:focus-visible) {
      :focus {
        outline: red solid 5px;
        background-color: red;
      }
    }

    :focus-visible {
      outline: red solid 5px;
    }

    :focus:not(:focus-visible) {
        outline: 0;
        background-color: lime;
    }
  </style>
</head>
<body>
  This test checks that focusing an element by a fragment navigation using a pointing device does not cause <code>:focus-visible</code> matching.
  <ol id="instructions">
    <li>If the user-agent does not claim to support the <code>:focus-visible</code> pseudo-class then SKIP this test.</li>
    <li>Click the link below that says "Click me"</li>
    <li>If the element that says "I will be focused by the fragment navigation" has a red outline, then the test result is FAILURE. If the element has a green background, then the test result is SUCCESS.</li>
  </ol>
  <br />

  <a id="link" href="#el" tabindex="0">Click me</a>
  <div id="el" tabindex="-1">I will be focused by the fragment navigation</div>

  <script>
    async_test(function(t) {
      el.addEventListener("focus", t.step_func(function() {
          assert_equals(getComputedStyle(el).backgroundColor, "rgb(0, 255, 0)", `backgroundColor for ${el.tagName}#${el.id} should be lime`);
          assert_not_equals(getComputedStyle(el).outlineColor, "rgb(255, 0, 0)", `outlineColor for ${el.tagName}#${el.id} should NOT be red`);
          t.done();
      }));
      test_driver.click(link);
    }, ":focus-visible should not match after a fragment navigation using a pointing device");
  </script>
</body>
</html>
